Refactor RegionFilter to have less duplicate code
