+++ aidants_connect/admin.py
@@ -1,6 +1,10 @@
+import operator
+from functools import reduce
+
 from admin_honeypot.admin import LoginAttemptAdmin as HoneypotLoginAttemptAdmin
 from admin_honeypot.models import LoginAttempt as HoneypotLoginAttempt
 from django.contrib.admin import SimpleListFilter
+from django.db.models import Q
 from django_celery_beat.admin import (
     ClockedScheduleAdmin,
     PeriodicTaskAdmin,
@@ -14,7 +18,11 @@ from django_celery_beat.models import (
 )
 from django_otp.admin import OTPAdminSite
 from magicauth.models import MagicToken
-from aidants_connect_web.models import DatavizRegion, DatavizDepartment
+from aidants_connect_web.models import (
+    DatavizRegion,
+    DatavizDepartment,
+    DatavizDepartmentsToRegion,
+)
 
 admin_site = OTPAdminSite(OTPAdminSite.name)
 admin_site.login_template = "aidants_connect_web/admin/login.html"
@@ -68,11 +76,36 @@ class RegionFilter(SimpleListFilter):
     title = "Région"
 
     parameter_name = "region"
+    filter_parameter_name = "zipcode"
 
     def lookups(self, request, model_admin):
         return [(r.id, r.name) for r in DatavizRegion.objects.all()] + [
             ("other", "Autre")
         ]
+
+    def queryset(self, request, queryset):
+        region_id = self.value()
+
+        if not region_id:
+            return
+
+        if region_id == "other":
+            return queryset.filter(**{self.filter_parameter_name: 0})
+
+        region = DatavizRegion.objects.get(id=region_id)
+        d2r = DatavizDepartmentsToRegion.objects.filter(region=region)
+        qgroup = reduce(
+            operator.or_,
+            (
+                Q(
+                    **{
+                        f"{self.filter_parameter_name}__startswith": d.department.normalize_zipcode()  # noqa: E501
+                    }
+                )
+                for d in d2r.all()
+            ),
+        )
+        return queryset.filter(qgroup)
 
 
 class DepartmentFilter(SimpleListFilter):
